Remove unused typedoc and rename files
