--- src/browser/components/SavedScripts/icons.tsx
@@ -1,6 +1,0 @@
-import React from 'react'
-import { Icon } from 'semantic-ui-react'
-
-export const ExpandMenuRightIcon = () => <Icon name="caret right" />
-
-export const CollapseMenuIcon = () => <Icon name="caret down" />
